Move MOverflowWrite from fold to VHF-common
Besides moving the relevant code, also implemented autoderef.
Thereafter fixed up some closure errors due to the migration.
