+++ VHF-rs/src/error.rs
@@ -61,6 +61,7 @@ impl From<vhf_common::Error> for Error {
             vhf_common::Error::Jiff(v) => Self::Jiff(v),
             vhf_common::Error::ParseEmpty => Self::ParseEmpty,
             vhf_common::Error::ParseUnrecognised(v) => Self::ParseUnrecognised(v),
+            vhf_common::Error::ExcessData => Self::ExcessData,
         }
     }
 }
